Revert optional imports in package init
