+++ functions/utility.py
@@ -322,3 +322,17 @@ def create_bad_records_table(settings, spark):
 
     if spark.catalog.tableExists(f"{dst_table_name}_bad_records"):
         raise Exception(f"Bad records table exists: {dst_table_name}_bad_records")
+
+
+def parse_si(value):
+    """Return ``value`` converted to a float using SI notation."""
+
+    if isinstance(value, (int, float)):
+        return float(value)
+    if not isinstance(value, str):
+        raise TypeError("value must be a number or string")
+    s = value.strip().lower()
+    scale = {"k": 1e3, "m": 1e6, "g": 1e9, "t": 1e12}
+    if s and s[-1] in scale:
+        return float(s[:-1]) * scale[s[-1]]
+    return float(s)
